requirements.txt
fastapi==0.116.1
uvicorn==0.29.0
requests==2.32.3
sentence-transformers==2.7.0
PyMuPDF==1.23.25
numpy==1.26.4
python-dotenv==1.0.1
google-generativeai==0.8.4
google-api-core==2.24.2
google-auth==2.38.0
